feat: authentication and authorization 
feat: routes with user authenticated in header request
feat: admin routes for sales/vehicle/sellers stats
TODO: error management

+++ src/main/java/com/develhope/spring/features/rentals/dto/PatchRentalRequest.java
@@ -1,12 +1,12 @@
 package com.develhope.spring.features.rentals.dto;
 
-import java.time.OffsetDateTime;
-
 import com.develhope.spring.features.users.UserEntity;
 import com.develhope.spring.features.vehicle.VehicleEntity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.OffsetDateTime;
 
 
 @Data
@@ -15,8 +15,8 @@ import lombok.NoArgsConstructor;
 public class PatchRentalRequest {
     private OffsetDateTime startOfRental;
     private OffsetDateTime endOfRental;
-    private Integer dailyCostRental;
-    private Integer totalCostRental;
+    private Long dailyCostRental;
+    private Long totalCostRental;
     private String paymentStatus;
     private VehicleEntity vehicleEntity;
     private UserEntity userEntity;
